feat(nvim): use custom git root detector
This adds a new `git` root detector to LazyVim's root util which
returns the most top level git directory as the root.

This solves the issue of opening a file which is in a subrepository
and having all the pickers now consider the subrepo directory
to be the root, making it very awkward to navigate or manipulate
any files in the actual root repository.

Idiomatic LazyVim usage is that this is intended behaviour, and
that the user should use (in pickers at least) the `cwd` toggle
which will use nvim's cwd (as reported by `vim.uv.cwd()`) as the
root instead.

However personally I find this a very grating decision as it does
not match my common workflows at all - where if I am looking at /
editing a file in a subrepo it is almost always to get more
context for / debug library behaviour, and I will often want to
quickly go back to a file in the actual project I am working on,
or a file in another library within that project - so this behaviour
only serves to cause me friction and frustration.

It is possible this change will have some unintended consequences
down the line, however I am currently unable to think of any which
could not be solved with some custom configuration for that case.

The only real issue I have encountered so far with this is that
for pickers it does not currently mesh well with my custom `libs`
toggle, as it hides those files/directories by default which can
be confusing if the root is hidden. My current plan to address this
is to figure out a way to set this toggle to be enabled by default
if the current root would be hidden by the toggle.

I also plan to change LazyVim's builtin `cwd` toggle behaviour to
toggle between the original root behaviour as I don't think the
original behaviour was entirely undesirable, just only as a default.
However I am a small bit concerned that mangling the `cwd` concept
could cause some issues.