refactor(nvim): compile fennel using nfnl through nvim
This should improve reproducibility by keeping the lua generated
as a build step in sync with files generated on saving a fennel
file in nvim.

The build script for the compiled lua derivation creates a temporary
home for running nvim, so config files can be loaded and a trustfile
can be created. It also now relies on nixvim as an input, which ensures
that all other config and plugins have been generated / loaded before
trying to compile the fennel.