feat(fish): add github PRs to prompt and fix async
Fixes issues with the how async prompts load, which was due to not
having the actual fish prompt function as an async one. I still don't
want it async (as I only want to set the actual async prompt subfunctions
for performance reasons), so I just manually copy it over in the way
that fish-refresh-prompt-on-cmd expects.

Github PRs are also added to the prompt in an async manner, along with
separating the diff info into it's own sub part and various performance
improvements for how the data is retrieved and cached.